requirements.txt
langgraph
langchain-google-genai
python-dotenv
fastapi
uvicorn[standard]
python-multipart
pypdf
python-docx
pillow
pytesseract

pydantic